rename: directory 이름 snake_case로 변경
- 플러터 기준에 맞춰 directory 이름들을 snake_case로 변경

- scriptInput -> script_inptut
- dataSource -> data_source
- themeInput -> theme_input
- voiceRecode -> voice_recode

#issue-11 directoryRename
