+++ lib/features/script_input/view/script_input_screen.dart
