+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_practice_mode_storage.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_practice_theme_storage.dart';
-import 'package:swm_peech_flutter/features/common/dataSource/local_script_storage.dart';
+import 'package:swm_peech_flutter/features/common/data_source/local_practice_mode_storage.dart';
+import 'package:swm_peech_flutter/features/common/data_source/local_practice_theme_storage.dart';
+import 'package:swm_peech_flutter/features/common/data_source/local_script_storage.dart';
 import 'package:swm_peech_flutter/routers/routers.dart';
 
 void main() {
